requirements.txt
blinker==1.9.0
Brotli==1.1.0
click==8.1.8
Flask==3.1.0
Flask-Compress==1.17
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
PyYAML==6.0.2
Werkzeug==3.1.3
zstandard==0.23.0
